+++ test/test_ul4on.py
@@ -201,6 +201,17 @@ def test_monthdelta(t):
 	assert d == t(d)
 
 
+def test_slice(t):
+	d = slice(None, None)
+	assert d == t(d)
+	d = slice(1, None)
+	assert d == t(d)
+	d = slice(None, 3)
+	assert d == t(d)
+	d = slice(1, 3)
+	assert d == t(d)
+
+
 def test_list(t):
 	assert [] == t([])
 	assert [1, 2, 3] == t([1, 2, 3])